Replace double with decimal for all game data numeric types (Task 12)
Task 12: doubledecimal conversion across the entire codebase

Models:
- ColonyStructureStatus: all 10 numeric fields (Power, Habitation, Food,
  Entertainment, Warehouse  provided/required) changed to decimal
- Item.Volume: double  decimal
- ItemProperty.BaseValue/AdjustedValue: double  decimal
- PropertyBag: getDouble renamed to getDecimal, setProperty(double)  setProperty(decimal)

Services:
- ColonyStatusCalculator: all accumulator variables, GetBlueprintDouble  GetBlueprintDecimal,
  AppendStatus parameters converted to decimal
- Colony.ProcessColony: extractionMultiplier, refiningMultiplier, quantity calculations
- ColonyBootstrap: miningRate, adjustedRate, refinedOutput, BestResourceEntry fields
- ColonyInactivityCollector: GetMiningOutputRate, totalMiningOutput, totalConsumption,
  supply, available all converted to decimal
- EvolutionChainService: SegmentInfo, EvolutionGraphData series, percentage calculations,
  ParseTimeToSeconds, GetNumericValue all converted to decimal
- BuildOrderOptimizer: all getDouble calls  getDecimal with decimal variables

Constants:
- GameConstants.WorkerVolume: double  decimal (removed unnecessary cast)

Forms:
- FormColony.GetItemVolume: return type double  decimal, all literals use m suffix
- ColonyStructure: power display variables double  decimal, getDouble  getDecimal

Tests:
- PropertyBagTests: getDouble  getDecimal, double  decimal assertions
- ColonyStructureTests: getDouble  getDecimal, double  decimal assertions
- EvolutionChainServiceTests: all double raw values  decimal, percentage comparisons
  use decimal, setProperty calls use decimal/m suffix
- EvolutionGraphRenderingTests: point generation uses decimal instead of double
- NEW: DecimalRoundTripPropertyTests (Property 6)  verifies decimal values
  survive JSON serialization round-trip with exact equality via Newtonsoft.Json

Sweep: zero remaining getDouble calls in the codebase (verified via grep).

All 1031 tests pass.

Spec: .kiro/specs/baseline-data-stability/
Prompt: Execute tasks 12.9 through 12.12 for the baseline-data-stability spec.

+++ OE2EmpireTracker.Tests/Forms/EvolutionGraphRenderingTests.cs
@@ -34,13 +34,13 @@ namespace OE2EmpireTracker.Tests.Forms
             {
                 // For each evolution level, generate a random percent value (1.0–500.0)
                 var percentsGen = Gen.Sequence(
-                    evLevels.Select(_ => Gen.Choose(1, 500).Select(v => (double)v))
+                    evLevels.Select(_ => Gen.Choose(1, 500).Select(v => (decimal)v))
                 );
 
                 return percentsGen.Select(percents =>
                 {
                     var pArr = percents.ToArray();
-                    var points = new List<(int Evolution, double Percent)>();
+                    var points = new List<(int Evolution, decimal Percent)>();
                     for (int i = 0; i < evLevels.Length; i++)
                     {
                         points.Add((evLevels[i], pArr[i]));
